+++ console/provider.go
@@ -114,6 +114,8 @@ func (p *ConsoleServiceProvider) Register(app contracts.Application) error {
 	p.kernel.AddCommand(commands.MakeViewCommand(app).Cobra(app))
 	p.kernel.AddCommand(commands.MakeComponentCommand(app).Cobra(app))
 	p.kernel.AddCommand(commands.MakeAuthCommand(app).Cobra(app))
+	p.kernel.AddCommand(commands.QueueMonitorCommand(app).Cobra(app))
+	p.kernel.AddCommand(commands.QueueRestartCommand(app).Cobra(app))
 
 	// Bind kernel to container
 	app.InstanceType(p.kernel)